+++ taiga/users/models.py
@@ -19,6 +19,7 @@ import os
 import os.path as path
 import random
 import re
+import uuid
 
 from unidecode import unidecode
 
@@ -33,6 +34,7 @@ from django.template.defaultfilters import slugify
 
 from djorm_pgarray.fields import TextArrayField
 
+from taiga.auth.tokens import get_token_for_user
 from taiga.base.utils.slug import slugify_uniquely
 from taiga.base.utils.iterators import split_by_n
 from taiga.permissions.permissions import MEMBERS_PERMISSIONS
@@ -152,6 +154,24 @@ class User(AbstractBaseUser, PermissionsMixin):
     def get_full_name(self):
         return self.full_name or self.username or self.email
 
+    def save(self, *args, **kwargs):
+        get_token_for_user(self, "cancel_account")
+        super().save(*args, **kwargs)
+
+    def cancel(self):
+        self.username = slugify_uniquely("deleted-user", User, slugfield="username")
+        self.email = "{}@taiga.io".format(self.username)
+        self.is_active = False
+        self.full_name = "Deleted user"
+        self.color = ""
+        self.bio = ""
+        self.default_language = ""
+        self.default_timezone = ""
+        self.colorize_tags = True
+        self.token = None
+        self.github_id = None
+        self.set_unusable_password()
+        self.save()
 
 class Role(models.Model):
     name = models.CharField(max_length=200, null=False, blank=False,